requirements.txt
fastapi
slowapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
alembic
python-dotenv
PyJWT
black
flake8
autoflake8
pytest
pytest-asyncio
pytest-env
httpx
isort
bcrypt